clean up periphery docker client
