+++ bin/periphery/src/docker/containers.rs
@@ -1,42 +1,15 @@
-use std::{collections::HashMap, sync::OnceLock};
+use std::collections::HashMap;
 
-use anyhow::{Context, anyhow};
-use bollard::{
-  Docker,
-  query_parameters::{
-    InspectContainerOptions, InspectNetworkOptions,
-    ListContainersOptions, ListImagesOptions, ListNetworksOptions,
-    ListVolumesOptions,
-  },
+use anyhow::Context;
+use bollard::query_parameters::{
+  InspectContainerOptions, ListContainersOptions,
 };
-use command::run_komodo_command;
-use komodo_client::entities::{
-  TerminationSignal,
-  docker::{
-    ContainerConfig, GraphDriverData, HealthConfig, PortBinding,
-    container::*, image::*, network::*, volume::*,
-  },
-  update::Log,
+use komodo_client::entities::docker::{
+  ContainerConfig, GraphDriverData, HealthConfig, PortBinding,
+  container::*,
 };
-use run_command::async_run_command;
 
-pub fn docker_client() -> &'static DockerClient {
-  static DOCKER_CLIENT: OnceLock<DockerClient> = OnceLock::new();
-  DOCKER_CLIENT.get_or_init(Default::default)
-}
-
-pub struct DockerClient {
-  docker: Docker,
-}
-
-impl Default for DockerClient {
-  fn default() -> DockerClient {
-    DockerClient {
-      docker: Docker::connect_with_defaults()
-        .expect("failed to connect to docker daemon"),
-    }
-  }
-}
+use super::DockerClient;
 
 impl DockerClient {
   pub async fn list_containers(
@@ -544,480 +517,5 @@ impl DockerClient {
         })).collect(),
     }),
     })
-  }
-
-  pub async fn list_networks(
-    &self,
-    containers: &[ContainerListItem],
-  ) -> anyhow::Result<Vec<NetworkListItem>> {
-    let networks = self
-      .docker
-      .list_networks(Option::<ListNetworksOptions>::None)
-      .await?
-      .into_iter()
-      .map(|network| {
-        let (ipam_driver, ipam_subnet, ipam_gateway) =
-          if let Some(ipam) = network.ipam {
-            let (subnet, gateway) = if let Some(config) = ipam
-              .config
-              .and_then(|configs| configs.into_iter().next())
-            {
-              (config.subnet, config.gateway)
-            } else {
-              (None, None)
-            };
-            (ipam.driver, subnet, gateway)
-          } else {
-            (None, None, None)
-          };
-        let in_use = match &network.name {
-          Some(name) => containers.iter().any(|container| {
-            container.networks.iter().any(|_name| name == _name)
-          }),
-          None => false,
-        };
-        NetworkListItem {
-          name: network.name,
-          id: network.id,
-          created: network.created,
-          scope: network.scope,
-          driver: network.driver,
-          enable_ipv6: network.enable_ipv6,
-          ipam_driver,
-          ipam_subnet,
-          ipam_gateway,
-          internal: network.internal,
-          attachable: network.attachable,
-          ingress: network.ingress,
-          in_use,
-        }
-      })
-      .collect();
-    Ok(networks)
-  }
-
-  pub async fn inspect_network(
-    &self,
-    network_name: &str,
-  ) -> anyhow::Result<Network> {
-    let network = self
-      .docker
-      .inspect_network(
-        network_name,
-        InspectNetworkOptions {
-          verbose: true,
-          ..Default::default()
-        }
-        .into(),
-      )
-      .await?;
-    Ok(Network {
-      name: network.name,
-      id: network.id,
-      created: network.created,
-      scope: network.scope,
-      driver: network.driver,
-      enable_ipv6: network.enable_ipv6,
-      ipam: network.ipam.map(|ipam| Ipam {
-        driver: ipam.driver,
-        config: ipam
-          .config
-          .unwrap_or_default()
-          .into_iter()
-          .map(|config| IpamConfig {
-            subnet: config.subnet,
-            ip_range: config.ip_range,
-            gateway: config.gateway,
-            auxiliary_addresses: config
-              .auxiliary_addresses
-              .unwrap_or_default(),
-          })
-          .collect(),
-        options: ipam.options.unwrap_or_default(),
-      }),
-      internal: network.internal,
-      attachable: network.attachable,
-      ingress: network.ingress,
-      containers: network
-        .containers
-        .unwrap_or_default()
-        .into_iter()
-        .map(|(container_id, container)| NetworkContainer {
-          container_id,
-          name: container.name,
-          endpoint_id: container.endpoint_id,
-          mac_address: container.mac_address,
-          ipv4_address: container.ipv4_address,
-          ipv6_address: container.ipv6_address,
-        })
-        .collect(),
-      options: network.options.unwrap_or_default(),
-      labels: network.labels.unwrap_or_default(),
-    })
-  }
-
-  pub async fn list_images(
-    &self,
-    containers: &[ContainerListItem],
-  ) -> anyhow::Result<Vec<ImageListItem>> {
-    let images = self
-      .docker
-      .list_images(Option::<ListImagesOptions>::None)
-      .await?
-      .into_iter()
-      .map(|image| {
-        let in_use = containers.iter().any(|container| {
-          container
-            .image_id
-            .as_ref()
-            .map(|id| id == &image.id)
-            .unwrap_or_default()
-        });
-        ImageListItem {
-          name: image
-            .repo_tags
-            .into_iter()
-            .next()
-            .unwrap_or_else(|| image.id.clone()),
-          id: image.id,
-          parent_id: image.parent_id,
-          created: image.created,
-          size: image.size,
-          in_use,
-        }
-      })
-      .collect();
-    Ok(images)
-  }
-
-  pub async fn inspect_image(
-    &self,
-    image_name: &str,
-  ) -> anyhow::Result<Image> {
-    let image = self.docker.inspect_image(image_name).await?;
-    Ok(Image {
-      id: image.id,
-      repo_tags: image.repo_tags.unwrap_or_default(),
-      repo_digests: image.repo_digests.unwrap_or_default(),
-      parent: image.parent,
-      comment: image.comment,
-      created: image.created,
-      docker_version: image.docker_version,
-      author: image.author,
-      architecture: image.architecture,
-      variant: image.variant,
-      os: image.os,
-      os_version: image.os_version,
-      size: image.size,
-      graph_driver: image.graph_driver.map(|driver| {
-        GraphDriverData {
-          name: driver.name,
-          data: driver.data,
-        }
-      }),
-      root_fs: image.root_fs.map(|fs| ImageInspectRootFs {
-        typ: fs.typ,
-        layers: fs.layers.unwrap_or_default(),
-      }),
-      metadata: image.metadata.map(|metadata| ImageInspectMetadata {
-        last_tag_time: metadata.last_tag_time,
-      }),
-      config: image.config.map(|config| ContainerConfig {
-        hostname: config.hostname,
-        domainname: config.domainname,
-        user: config.user,
-        attach_stdin: config.attach_stdin,
-        attach_stdout: config.attach_stdout,
-        attach_stderr: config.attach_stderr,
-        exposed_ports: config
-          .exposed_ports
-          .unwrap_or_default()
-          .into_keys()
-          .map(|k| (k, Default::default()))
-          .collect(),
-        tty: config.tty,
-        open_stdin: config.open_stdin,
-        stdin_once: config.stdin_once,
-        env: config.env.unwrap_or_default(),
-        cmd: config.cmd.unwrap_or_default(),
-        healthcheck: config.healthcheck.map(|health| HealthConfig {
-          test: health.test.unwrap_or_default(),
-          interval: health.interval,
-          timeout: health.timeout,
-          retries: health.retries,
-          start_period: health.start_period,
-          start_interval: health.start_interval,
-        }),
-        args_escaped: config.args_escaped,
-        image: config.image,
-        volumes: config
-          .volumes
-          .unwrap_or_default()
-          .into_keys()
-          .map(|k| (k, Default::default()))
-          .collect(),
-        working_dir: config.working_dir,
-        entrypoint: config.entrypoint.unwrap_or_default(),
-        network_disabled: config.network_disabled,
-        mac_address: config.mac_address,
-        on_build: config.on_build.unwrap_or_default(),
-        labels: config.labels.unwrap_or_default(),
-        stop_signal: config.stop_signal,
-        stop_timeout: config.stop_timeout,
-        shell: config.shell.unwrap_or_default(),
-      }),
-    })
-  }
-
-  pub async fn image_history(
-    &self,
-    image_name: &str,
-  ) -> anyhow::Result<Vec<ImageHistoryResponseItem>> {
-    let res = self
-      .docker
-      .image_history(image_name)
-      .await?
-      .into_iter()
-      .map(|image| ImageHistoryResponseItem {
-        id: image.id,
-        created: image.created,
-        created_by: image.created_by,
-        tags: image.tags,
-        size: image.size,
-        comment: image.comment,
-      })
-      .collect();
-    Ok(res)
-  }
-
-  pub async fn list_volumes(
-    &self,
-    containers: &[ContainerListItem],
-  ) -> anyhow::Result<Vec<VolumeListItem>> {
-    let volumes = self
-      .docker
-      .list_volumes(Option::<ListVolumesOptions>::None)
-      .await?
-      .volumes
-      .unwrap_or_default()
-      .into_iter()
-      .map(|volume| {
-        let scope = volume
-          .scope
-          .map(|scope| match scope {
-            bollard::secret::VolumeScopeEnum::EMPTY => {
-              VolumeScopeEnum::Empty
-            }
-            bollard::secret::VolumeScopeEnum::LOCAL => {
-              VolumeScopeEnum::Local
-            }
-            bollard::secret::VolumeScopeEnum::GLOBAL => {
-              VolumeScopeEnum::Global
-            }
-          })
-          .unwrap_or(VolumeScopeEnum::Empty);
-        let in_use = containers.iter().any(|container| {
-          container.volumes.iter().any(|name| &volume.name == name)
-        });
-        VolumeListItem {
-          name: volume.name,
-          driver: volume.driver,
-          mountpoint: volume.mountpoint,
-          created: volume.created_at,
-          size: volume.usage_data.map(|data| data.size),
-          scope,
-          in_use,
-        }
-      })
-      .collect();
-    Ok(volumes)
-  }
-
-  pub async fn inspect_volume(
-    &self,
-    volume_name: &str,
-  ) -> anyhow::Result<Volume> {
-    let volume = self.docker.inspect_volume(volume_name).await?;
-    Ok(Volume {
-      name: volume.name,
-      driver: volume.driver,
-      mountpoint: volume.mountpoint,
-      created_at: volume.created_at,
-      status: volume.status.unwrap_or_default().into_keys().map(|k| (k, Default::default())).collect(),
-      labels: volume.labels,
-      scope: volume
-        .scope
-        .map(|scope| match scope {
-          bollard::secret::VolumeScopeEnum::EMPTY => {
-            VolumeScopeEnum::Empty
-          }
-          bollard::secret::VolumeScopeEnum::LOCAL => {
-            VolumeScopeEnum::Local
-          }
-          bollard::secret::VolumeScopeEnum::GLOBAL => {
-            VolumeScopeEnum::Global
-          }
-        })
-        .unwrap_or_default(),
-      cluster_volume: volume.cluster_volume.map(|volume| {
-        ClusterVolume {
-          id: volume.id,
-          version: volume.version.map(|version| ObjectVersion {
-            index: version.index,
-          }),
-          created_at: volume.created_at,
-          updated_at: volume.updated_at,
-          spec: volume.spec.map(|spec| ClusterVolumeSpec {
-            group: spec.group,
-            access_mode: spec.access_mode.map(|mode| {
-              ClusterVolumeSpecAccessMode {
-                scope: mode.scope.map(|scope| match scope {
-                  bollard::secret::ClusterVolumeSpecAccessModeScopeEnum::EMPTY => ClusterVolumeSpecAccessModeScopeEnum::Empty,
-                  bollard::secret::ClusterVolumeSpecAccessModeScopeEnum::SINGLE => ClusterVolumeSpecAccessModeScopeEnum::Single,
-                  bollard::secret::ClusterVolumeSpecAccessModeScopeEnum::MULTI => ClusterVolumeSpecAccessModeScopeEnum::Multi,
-                }).unwrap_or_default(),
-                sharing: mode.sharing.map(|sharing| match sharing {
-                  bollard::secret::ClusterVolumeSpecAccessModeSharingEnum::EMPTY => ClusterVolumeSpecAccessModeSharingEnum::Empty,
-                  bollard::secret::ClusterVolumeSpecAccessModeSharingEnum::NONE => ClusterVolumeSpecAccessModeSharingEnum::None,
-                  bollard::secret::ClusterVolumeSpecAccessModeSharingEnum::READONLY => ClusterVolumeSpecAccessModeSharingEnum::Readonly,
-                  bollard::secret::ClusterVolumeSpecAccessModeSharingEnum::ONEWRITER => ClusterVolumeSpecAccessModeSharingEnum::Onewriter,
-                  bollard::secret::ClusterVolumeSpecAccessModeSharingEnum::ALL => ClusterVolumeSpecAccessModeSharingEnum::All,
-                }).unwrap_or_default(),
-                secrets: mode.secrets.unwrap_or_default().into_iter().map(|secret| ClusterVolumeSpecAccessModeSecrets {
-                    key: secret.key,
-                    secret: secret.secret,
-                }).collect(),
-                accessibility_requirements: mode
-                  .accessibility_requirements.map(|req| ClusterVolumeSpecAccessModeAccessibilityRequirements {
-                    requisite: req.requisite.unwrap_or_default().into_iter().map(|map| map.into_iter().map(|(k, v)| (k, v.unwrap_or_default().into_iter().map(|p| PortBinding { host_ip: p.host_ip, host_port: p.host_port }).collect())).collect()).collect(),
-                    preferred: req.preferred.unwrap_or_default().into_iter().map(|map| map.into_iter().map(|(k, v)| (k, v.unwrap_or_default().into_iter().map(|p| PortBinding { host_ip: p.host_ip, host_port: p.host_port }).collect())).collect()).collect(),
-                }),
-                capacity_range: mode.capacity_range.map(|range| ClusterVolumeSpecAccessModeCapacityRange {
-                  required_bytes: range.required_bytes,
-                  limit_bytes: range.limit_bytes,
-                }),
-                availability: mode.availability.map(|availability| match availability {
-                  bollard::secret::ClusterVolumeSpecAccessModeAvailabilityEnum::EMPTY => ClusterVolumeSpecAccessModeAvailabilityEnum::Empty,
-                  bollard::secret::ClusterVolumeSpecAccessModeAvailabilityEnum::ACTIVE => ClusterVolumeSpecAccessModeAvailabilityEnum::Active,
-                  bollard::secret::ClusterVolumeSpecAccessModeAvailabilityEnum::PAUSE => ClusterVolumeSpecAccessModeAvailabilityEnum::Pause,
-                  bollard::secret::ClusterVolumeSpecAccessModeAvailabilityEnum::DRAIN => ClusterVolumeSpecAccessModeAvailabilityEnum::Drain,
-                }).unwrap_or_default(),
-              }
-            }),
-          }),
-          info: volume.info.map(|info| ClusterVolumeInfo {
-            capacity_bytes: info.capacity_bytes,
-            volume_context: info.volume_context.unwrap_or_default(),
-            volume_id: info.volume_id,
-            accessible_topology: info.accessible_topology.unwrap_or_default().into_iter().map(|map| map.into_iter().map(|(k, v)| (k, v.unwrap_or_default().into_iter().map(|p| PortBinding { host_ip: p.host_ip, host_port: p.host_port }).collect())).collect()).collect(),
-          }),
-          publish_status: volume
-            .publish_status
-            .unwrap_or_default()
-            .into_iter()
-            .map(|status| ClusterVolumePublishStatus {
-              node_id: status.node_id,
-              state: status.state.map(|state| match state {
-                bollard::secret::ClusterVolumePublishStatusStateEnum::EMPTY => ClusterVolumePublishStatusStateEnum::Empty,
-                bollard::secret::ClusterVolumePublishStatusStateEnum::PENDING_PUBLISH => ClusterVolumePublishStatusStateEnum::PendingPublish,
-                bollard::secret::ClusterVolumePublishStatusStateEnum::PUBLISHED => ClusterVolumePublishStatusStateEnum::Published,
-                bollard::secret::ClusterVolumePublishStatusStateEnum::PENDING_NODE_UNPUBLISH => ClusterVolumePublishStatusStateEnum::PendingNodeUnpublish,
-                bollard::secret::ClusterVolumePublishStatusStateEnum::PENDING_CONTROLLER_UNPUBLISH => ClusterVolumePublishStatusStateEnum::PendingControllerUnpublish,
-              }).unwrap_or_default(),
-              publish_context: status.publish_context.unwrap_or_default(),
-            })
-            .collect(),
-        }
-      }),
-      options: volume.options,
-      usage_data: volume.usage_data.map(|data| VolumeUsageData {
-        size: data.size,
-        ref_count: data.ref_count,
-      }),
-    })
-  }
-}
-
-/// Returns whether build result should be pushed after build
-#[instrument(skip(registry_token))]
-pub async fn docker_login(
-  domain: &str,
-  account: &str,
-  // For local token override from core.
-  registry_token: Option<&str>,
-) -> anyhow::Result<bool> {
-  if domain.is_empty() || account.is_empty() {
-    return Ok(false);
-  }
-  let registry_token = match registry_token {
-    Some(token) => token,
-    None => crate::helpers::registry_token(domain, account)?,
-  };
-  let log = async_run_command(&format!(
-    "echo {registry_token} | docker login {domain} --username '{account}' --password-stdin",
-  ))
-  .await;
-  if log.success() {
-    Ok(true)
-  } else {
-    let mut e = anyhow!("End of trace");
-    for line in
-      log.stderr.split('\n').filter(|line| !line.is_empty()).rev()
-    {
-      e = e.context(line.to_string());
-    }
-    for line in
-      log.stdout.split('\n').filter(|line| !line.is_empty()).rev()
-    {
-      e = e.context(line.to_string());
-    }
-    Err(e.context(format!("Registry {domain} login error")))
-  }
-}
-
-#[instrument]
-pub async fn pull_image(image: &str) -> Log {
-  let command = format!("docker pull {image}");
-  run_komodo_command("Docker Pull", None, command).await
-}
-
-pub fn stop_container_command(
-  container_name: &str,
-  signal: Option<TerminationSignal>,
-  time: Option<i32>,
-) -> String {
-  let signal = signal
-    .map(|signal| format!(" --signal {signal}"))
-    .unwrap_or_default();
-  let time = time
-    .map(|time| format!(" --time {time}"))
-    .unwrap_or_default();
-  format!("docker stop{signal}{time} {container_name}")
-}
-
-pub async fn container_stats(
-  container_name: Option<String>,
-) -> anyhow::Result<Vec<ContainerStats>> {
-  let format = "--format \"{{ json . }}\"";
-  let container_name = match container_name {
-    Some(name) => format!(" {name}"),
-    None => "".to_string(),
-  };
-  let command =
-    format!("docker stats{container_name} --no-stream {format}");
-  let output = async_run_command(&command).await;
-  if output.success() {
-    output
-      .stdout
-      .split('\n')
-      .filter(|e| !e.is_empty())
-      .map(|e| {
-        let parsed = serde_json::from_str(e)
-          .context(format!("failed at parsing entry {e}"))?;
-        Ok(parsed)
-      })
-      .collect()
-  } else {
-    Err(anyhow!("{}", output.stderr.replace('\n', "")))
   }
 }
